Design and Coding for the Main Activity
Project for: Android 5.0
Type: ADD | MODIFIED
Datetime: 19:33 23-01-2017

Design of the MainActivty, which contains the principal function of
WearApp project. The high, perfect and low contrast are getting them.
The future work still remains in the adjustement of the variables whose
are regulating the average color.

The landscape activities Main and Splash are ready.